+++ client/Lykke.Service.HFT.Wamp.Client/Lykke.Service.HFT.Wamp.Client.csproj
@@ -13,4 +13,8 @@
     <PackageReference Include="Lykke.WampSharp.Default.Client" Version="1.0.2" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\Lykke.Service.HFT.Contracts\Lykke.Service.HFT.Contracts.csproj" />
+  </ItemGroup>
+
 </Project>
